raspi/requirements.txt
pyserial
pyserial-asyncio
aiosqlite
pandas
scikit-learn
joblib
numpy
schedule
python-telegram-bot
